fix(db): separate current state of the deployments from history
Ref: SRX-5WTFDM
