+++ pkg/db/db.go
@@ -928,30 +928,6 @@ func (h *DBHandler) DBSelectAllApplications(ctx context.Context, transaction *sq
 	return &resultGo, nil
 }
 
-type DBDeployment struct {
-	Created        time.Time
-	ReleaseVersion *int64
-	App            string
-	Env            string
-	TransformerID  TransformerID
-	Metadata       string // json
-}
-
-type Deployment struct {
-	Created       time.Time
-	App           string
-	Env           string
-	Version       *int64
-	Metadata      DeploymentMetadata
-	TransformerID TransformerID
-}
-
-type DeploymentMetadata struct {
-	DeployedByName  string
-	DeployedByEmail string
-	CiLink          string
-}
-
 type EnvironmentLock struct {
 	EslVersion EslVersion
 	Created    time.Time
@@ -1026,7 +1002,6 @@ func (h *DBHandler) RunCustomMigrations(
 	ctx context.Context,
 	getAllAppsFun GetAllAppsFun,
 	writeAllDeploymentsFun WriteAllDeploymentsFun,
-	writeAllCurrentlyDeployedFun WriteAllDeploymentsFun,
 	writeAllReleasesFun WriteAllReleasesFun,
 	writeAllEnvLocksFun WriteAllEnvLocksFun,
 	writeAllAppLocksFun WriteAllAppLocksFun,
@@ -1046,10 +1021,6 @@ func (h *DBHandler) RunCustomMigrations(
 		return err
 	}
 	err = h.RunCustomMigrationDeployments(ctx, writeAllDeploymentsFun)
-	if err != nil {
-		return err
-	}
-	err = h.RunCustomMigrationAllDeployments(ctx, writeAllCurrentlyDeployedFun) //TODO: Merge with RunCustomMigrationDeployments
 	if err != nil {
 		return err
 	}
@@ -1078,440 +1049,6 @@ func (h *DBHandler) RunCustomMigrations(
 		return err // better wrap the error in a descriptive message?
 	}
 	return nil
-}
-
-func (h *DBHandler) DBSelectLatestDeployment(ctx context.Context, tx *sql.Tx, appSelector string, envSelector string) (*Deployment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectLatestDeployment")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, releaseVersion, appName, envName, metadata, transformereslVersion" +
-			" FROM deployments " +
-			" WHERE appName=? AND envName=? " +
-			" ORDER BY version DESC " +
-			" LIMIT 1;"))
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		appSelector,
-		envSelector,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not select deployment from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("deployments: row closing error: %v", err)
-		}
-	}(rows)
-	return processDeployment(rows)
-}
-
-func (h *DBHandler) DBSelectAllLatestDeploymentsForApplication(ctx context.Context, tx *sql.Tx, appName string) (map[string]Deployment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllLatestDeployments")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(
-		`
-	SELECT
-		deployments.created,
-		deployments.appname,
-		deployments.releaseVersion,
-		deployments.envName,
-		deployments.metadata
-	FROM (
-	SELECT
-		MAX(version) AS latest,
-		appname,
-		envname
-	FROM
-		deployments
-	GROUP BY
-		envName, appname
-	) AS latest
-	JOIN
-		deployments AS deployments
-	ON
-		latest.latest=deployments.version
-		AND latest.appname=deployments.appname
-		AND latest.envName=deployments.envName
-	WHERE deployments.appname = (?) AND deployments.releaseVersion IS NOT NULL ;`)
-
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		appName,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not select deployment from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("deployments: row closing error: %v", err)
-		}
-	}(rows)
-	return processAllLatestDeploymentsForApp(rows)
-}
-
-func (h *DBHandler) DBSelectAllLatestDeploymentsOnEnvironment(ctx context.Context, tx *sql.Tx, envName string) (map[string]*int64, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllLatestDeployments")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(
-		`SELECT
- deployments.appname,
- deployments.releaseVersion
-FROM (
- SELECT
-   MAX(version) AS latest,
-   appname,
-   envname
- FROM
-   deployments
- GROUP BY
-   envName, appname) AS latest
-JOIN
- deployments AS deployments
-ON
- latest.latest=deployments.version
- AND latest.appname=deployments.appname
- AND latest.envName=deployments.envName
-WHERE deployments.envName= ?;
-`)
-
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		envName,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not select deployment from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("deployments: row closing error: %v", err)
-		}
-	}(rows)
-	return processAllLatestDeployments(rows)
-}
-
-func processAllLatestDeployments(rows *sql.Rows) (map[string]*int64, error) {
-	result := make(map[string]*int64)
-	for rows.Next() {
-		var releaseVersion sql.NullInt64
-		var appName string
-		err := rows.Scan(&appName, &releaseVersion)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning deployments row from DB. Error: %w\n", err)
-		}
-
-		if releaseVersion.Valid {
-			result[appName] = &releaseVersion.Int64
-		}
-	}
-	err := rows.Close()
-	if err != nil {
-		return nil, fmt.Errorf("deployments: row closing error: %v\n", err)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, fmt.Errorf("deployments: row has error: %v\n", err)
-	}
-	return result, nil
-}
-
-func processAllLatestDeploymentsForApp(rows *sql.Rows) (map[string]Deployment, error) {
-	result := make(map[string]Deployment)
-	for rows.Next() {
-		var curr = Deployment{
-			Created: time.Time{},
-			Env:     "",
-			App:     "",
-			Version: nil,
-			Metadata: DeploymentMetadata{
-				DeployedByName:  "",
-				DeployedByEmail: "",
-				CiLink:          "",
-			},
-			TransformerID: 0,
-		}
-		var releaseVersion sql.NullInt64
-		var jsonMetadata string
-		err := rows.Scan(&curr.Created, &curr.App, &releaseVersion, &curr.Env, &jsonMetadata)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning deployments row from DB. Error: %w\n", err)
-		}
-		err = json.Unmarshal(([]byte)(jsonMetadata), &curr.Metadata)
-		if err != nil {
-			return nil, fmt.Errorf("Error during json unmarshal in deployments. Error: %w. Data: %s\n", err, jsonMetadata)
-		}
-		if releaseVersion.Valid {
-			curr.Version = &releaseVersion.Int64
-		}
-		result[curr.Env] = curr
-	}
-	err := rows.Close()
-	if err != nil {
-		return nil, fmt.Errorf("deployments: row closing error: %v\n", err)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, fmt.Errorf("deployments: row has error: %v\n", err)
-	}
-	return result, nil
-}
-
-func (h *DBHandler) DBSelectSpecificDeployment(ctx context.Context, tx *sql.Tx, appSelector string, envSelector string, releaseVersion uint64) (*Deployment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectSpecificDeployment")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, releaseVersion, appName, envName, metadata, transformereslVersion" +
-			" FROM deployments " +
-			" WHERE appName=? AND envName=? and releaseVersion=?" +
-			" ORDER BY version DESC " +
-			" LIMIT 1;"))
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		appSelector,
-		envSelector,
-		releaseVersion,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not select deployment from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("deployments: row closing error: %v", err)
-		}
-	}(rows)
-	return processDeployment(rows)
-}
-
-func processDeployment(rows *sql.Rows) (*Deployment, error) {
-	var releaseVersion sql.NullInt64
-	var row = &DBDeployment{
-		Created:        time.Time{},
-		ReleaseVersion: nil,
-		App:            "",
-		Env:            "",
-		Metadata:       "",
-		TransformerID:  0,
-	}
-	//exhaustruct:ignore
-	var resultJson = DeploymentMetadata{}
-	if rows.Next() {
-		err := rows.Scan(&row.Created, &releaseVersion, &row.App, &row.Env, &row.Metadata, &row.TransformerID)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning deployments row from DB. Error: %w\n", err)
-		}
-		if releaseVersion.Valid {
-			row.ReleaseVersion = &releaseVersion.Int64
-		}
-
-		err = json.Unmarshal(([]byte)(row.Metadata), &resultJson)
-		if err != nil {
-			return nil, fmt.Errorf("Error during json unmarshal in deployments. Error: %w. Data: %s\n", err, row.Metadata)
-		}
-	}
-	err := rows.Close()
-	if err != nil {
-		return nil, fmt.Errorf("deployments: row closing error: %v\n", err)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, fmt.Errorf("deployments: row has error: %v\n", err)
-	}
-	return &Deployment{
-		Created:       row.Created,
-		App:           row.App,
-		Env:           row.Env,
-		Version:       row.ReleaseVersion,
-		Metadata:      resultJson,
-		TransformerID: row.TransformerID,
-	}, nil
-}
-
-func (h *DBHandler) DBSelectDeploymentHistory(ctx context.Context, tx *sql.Tx, appSelector string, envSelector string, limit int) ([]Deployment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectDeploymentHistory")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, releaseVersion, appName, envName, metadata, transformereslVersion" +
-			" FROM deployments " +
-			" WHERE appName=? AND envName=? " +
-			" ORDER BY version DESC " +
-			" LIMIT ?;"))
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		appSelector,
-		envSelector,
-		limit,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not select deployment history from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("deployments: row closing error: %v", err)
-		}
-	}(rows)
-
-	result := make([]Deployment, 0)
-
-	for rows.Next() {
-		row, err := h.processSingleDeploymentRow(ctx, rows)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, *row)
-	}
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-func (h *DBHandler) DBSelectDeploymentsByTransformerID(ctx context.Context, tx *sql.Tx, transformerID TransformerID, limit uint) ([]Deployment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectDeploymentsByTransformerID")
-	defer span.Finish()
-
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, releaseVersion, appName, envName, metadata, transformereslVersion" +
-			" FROM deployments " +
-			" WHERE transformereslVersion=? " +
-			" ORDER BY version DESC " +
-			" LIMIT ?;"))
-
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		transformerID,
-		limit,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not select deployments by transformer id from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("deployments: row closing error: %v", err)
-		}
-	}(rows)
-	deployments := make([]Deployment, 0)
-	for rows.Next() {
-		row, err := h.processSingleDeploymentRow(ctx, rows)
-		if err != nil {
-			return nil, err
-		}
-		deployments = append(deployments, *row)
-	}
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return deployments, nil
-}
-
-func (h *DBHandler) DBSelectAnyDeployment(ctx context.Context, tx *sql.Tx) (*DBDeployment, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAnyDeployment")
-	defer span.Finish()
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, releaseVersion, appName, envName" +
-			" FROM deployments " +
-			" LIMIT 1;"))
-	span.SetTag("query", selectQuery)
-
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not select any deployments from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("deployments row could not be closed: %v", err)
-		}
-	}(rows)
-	//exhaustruct:ignore
-	var row = &DBDeployment{}
-	if rows.Next() {
-		var releaseVersion sql.NullInt64
-		err := rows.Scan(&row.Created, &releaseVersion, &row.App, &row.Env)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning deployments row from DB. Error: %w\n", err)
-		}
-		if releaseVersion.Valid {
-			row.ReleaseVersion = &releaseVersion.Int64
-		}
-	} else {
-		row = nil
-	}
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return row, nil
-}
-
-func (h *DBHandler) DBAllDeploymentsContainsData(ctx context.Context, tx *sql.Tx) (bool, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBAllDeploymentsContainsData")
-	defer span.Finish()
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT eslVersion, created, appName, json" +
-			" FROM all_deployments " +
-			" LIMIT 1;"))
-	span.SetTag("query", selectQuery)
-
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-	)
-	if err != nil {
-		return false, fmt.Errorf("could not select any all_deployments from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("all_deployments row could not be closed: %v", err)
-		}
-	}(rows)
-	//exhaustruct:ignore
-	result := rows.Next()
-
-	err = closeRows(rows)
-	if err != nil {
-		return false, err
-	}
-	return result, nil
 }
 
 type DBApp struct {
@@ -1783,169 +1320,6 @@ func processAppRow(ctx context.Context, rows *sql.Rows) (*DBAppWithMetaData, err
 	return row, nil
 }
 
-// DBWriteDeployment writes one deployment, meaning "what should be deployed"
-func (h *DBHandler) DBWriteDeployment(ctx context.Context, tx *sql.Tx, deployment Deployment, skipOverview bool) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBWriteDeployment")
-	defer span.Finish()
-	if h == nil {
-		return nil
-	}
-	if tx == nil {
-		return fmt.Errorf("DBWriteDeployment: no transaction provided")
-	}
-
-	jsonToInsert, err := json.Marshal(deployment.Metadata)
-	if err != nil {
-		return fmt.Errorf("could not marshal json data: %w", err)
-	}
-
-	insertQuery := h.AdaptQuery(
-		"INSERT INTO deployments (created, releaseVersion, appName, envName, metadata, transformereslVersion) VALUES (?, ?, ?, ?, ?, ?);")
-
-	now, err := h.DBReadTransactionTimestamp(ctx, tx)
-	if err != nil {
-		return fmt.Errorf("DBWriteDeployment unable to get transaction timestamp: %w", err)
-	}
-	span.SetTag("query", insertQuery)
-	nullVersion := NewNullInt(deployment.Version)
-
-	_, err = tx.Exec(
-		insertQuery,
-		*now,
-		nullVersion,
-		deployment.App,
-		deployment.Env,
-		jsonToInsert,
-		deployment.TransformerID)
-
-	if err != nil {
-		return fmt.Errorf("could not write deployment into DB. Error: %w\n", err)
-	}
-	return nil
-}
-
-// DBSelectAllDeploymentsForApp Returns most recent version of deployments for app with name 'appName'
-func (h *DBHandler) DBSelectAllDeploymentsForApp(ctx context.Context, tx *sql.Tx, appName string) (*AllDeploymentsForApp, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllDeploymentsForApp")
-	defer span.Finish()
-	if h == nil {
-		return nil, nil
-	}
-	if tx == nil {
-		return nil, fmt.Errorf("DBSelectAllDeploymentsForApp: no transaction provided")
-	}
-
-	insertQuery := h.AdaptQuery(
-		"SELECT eslVersion, created, appName, json FROM all_deployments WHERE appName = (?) ORDER BY eslVersion DESC LIMIT 1;")
-
-	span.SetTag("query", insertQuery)
-	rows, err := tx.Query(
-		insertQuery,
-		appName,
-	)
-
-	return h.processAllDeploymentRow(ctx, err, rows)
-}
-
-// DBSelectAllDeploymentsForAppAtTimestamp Returns most recent version of deployments for app with name 'appName' at timestamp ts
-func (h *DBHandler) DBSelectAllDeploymentsForAppAtTimestamp(ctx context.Context, tx *sql.Tx, appName string, ts time.Time) (*AllDeploymentsForApp, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllDeploymentsForAppAtTimestamp")
-	defer span.Finish()
-	if h == nil {
-		return nil, nil
-	}
-	if tx == nil {
-		return nil, fmt.Errorf("DBSelectAllDeploymentsForAppAtTimestamp: no transaction provided")
-	}
-
-	insertQuery := h.AdaptQuery(
-		"SELECT eslVersion, created, appName, json " +
-			"FROM all_deployments " +
-			"WHERE appName = (?) AND created <= (?) " +
-			"ORDER BY eslVersion " +
-			"DESC LIMIT 1;")
-
-	span.SetTag("query", insertQuery)
-	rows, err := tx.Query(
-		insertQuery,
-		appName,
-		ts,
-	)
-
-	return h.processAllDeploymentRow(ctx, err, rows)
-}
-
-// DBUpdateAllDeploymentsForApp Updates table entry for application with name 'appName' with 'envName' -> 'version'
-func (h *DBHandler) DBUpdateAllDeploymentsForApp(ctx context.Context, tx *sql.Tx, appName, envName string, version int64) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBUpdateAllDeploymentsForApp")
-	defer span.Finish()
-	if h == nil {
-		return nil
-	}
-	if tx == nil {
-		return fmt.Errorf("DBUpdateAllDeploymentsForApp: no transaction provided")
-	}
-
-	//Get current deployment information for app
-	currentAllDeployments, err := h.DBSelectAllDeploymentsForApp(ctx, tx, appName)
-	if err != nil {
-		return fmt.Errorf("could not read current all deployments for app: '%s': %v", appName, err)
-	}
-
-	//Update current deployment layout with envName -> version
-	var deploymentsMap map[string]int64
-	var previousVersion EslVersion
-	if currentAllDeployments == nil { //New app
-		deploymentsMap = map[string]int64{}
-		previousVersion = 0
-	} else {
-		deploymentsMap = currentAllDeployments.Deployments
-		previousVersion = currentAllDeployments.Version
-
-	}
-	deploymentsMap[envName] = version
-
-	//Insert new information into the db
-	err = h.DBWriteAllDeploymentsForApp(ctx, tx, int(previousVersion), appName, deploymentsMap)
-	if err != nil {
-		return fmt.Errorf("could not write all deployments for app: '%s': %v", appName, err)
-	}
-	return nil
-}
-
-// DBClearAllDeploymentsForApp Clears all deployments map. Used when undeploying an application
-func (h *DBHandler) DBClearAllDeploymentsForApp(ctx context.Context, tx *sql.Tx, appName string) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBUpdateAllDeploymentsForApp")
-	defer span.Finish()
-	if h == nil {
-		return nil
-	}
-	if tx == nil {
-		return fmt.Errorf("DBClearAllDeploymentsForApp: no transaction provided")
-	}
-
-	//Get current deployment information for app
-	currentAllDeployments, err := h.DBSelectAllDeploymentsForApp(ctx, tx, appName)
-	if err != nil {
-		return fmt.Errorf("could not read current all deployments for app: '%s': %v", appName, err)
-	}
-
-	deploymentsMap := map[string]int64{} //EmptyMap
-	var previousVersion EslVersion
-	if currentAllDeployments == nil {
-		previousVersion = 0
-	} else {
-		previousVersion = currentAllDeployments.Version
-	}
-
-	//Insert new information into the db
-	err = h.DBWriteAllDeploymentsForApp(ctx, tx, int(previousVersion), appName, deploymentsMap)
-	if err != nil {
-		return fmt.Errorf("could not write all deployments for app: '%s': %v", appName, err)
-	}
-	return nil
-}
-
 // CUSTOM MIGRATIONS
 
 func (h *DBHandler) RunCustomMigrationReleases(ctx context.Context, getAllAppsFun GetAllAppsFun, writeAllReleasesFun WriteAllReleasesFun) error {
@@ -2013,27 +1387,6 @@ func (h *DBHandler) RunCustomMigrationDeployments(ctx context.Context, getAllDep
 	})
 }
 
-func (h *DBHandler) RunCustomMigrationAllDeployments(ctx context.Context, writeAllDeployments WriteAllDeploymentsFun) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "RunCustomMigrationDeployments")
-	defer span.Finish()
-
-	return h.WithTransaction(ctx, false, func(ctx context.Context, transaction *sql.Tx) error {
-		needsMigrating, err := h.needsAllDeploymentsMigrations(ctx, transaction)
-		if err != nil {
-			return err
-		}
-		if !needsMigrating {
-			return nil
-		}
-		err = writeAllDeployments(ctx, transaction, h)
-		if err != nil {
-			return fmt.Errorf("could not get current deployments to run custom migrations: %v", err)
-		}
-
-		return nil
-	})
-}
-
 func (h *DBHandler) needsDeploymentsMigrations(ctx context.Context, transaction *sql.Tx) (bool, error) {
 	l := logger.FromContext(ctx).Sugar()
 	allAppsDb, err := h.DBSelectAnyDeployment(ctx, transaction)
@@ -2045,19 +1398,6 @@ func (h *DBHandler) needsDeploymentsMigrations(ctx context.Context, transaction 
 		return false, nil
 	}
 	return true, nil
-}
-
-func (h *DBHandler) needsAllDeploymentsMigrations(ctx context.Context, transaction *sql.Tx) (bool, error) {
-	l := logger.FromContext(ctx).Sugar()
-	contains, err := h.DBAllDeploymentsContainsData(ctx, transaction)
-	if err != nil {
-		return true, err
-	}
-	if contains {
-		l.Warnf("There are already all_deployments in the DB - skipping migrations")
-		return !contains, nil
-	}
-	return !contains, nil
 }
 
 type AllApplicationsJson struct {
@@ -2239,7 +1579,6 @@ func (h *DBHandler) NeedsMigrations(ctx context.Context) (bool, error) {
 			(*DBHandler).needsTeamLocksMigrations,
 			(*DBHandler).needsCommitEventsMigrations,
 			(*DBHandler).needsEnvironmentsMigrations,
-			(*DBHandler).needsAllDeploymentsMigrations,
 		}
 		for i := range checkFunctions {
 			f := checkFunctions[i]
@@ -4559,46 +3898,6 @@ func (h *DBHandler) processSingleDeploymentAttemptsRow(ctx context.Context, rows
 
 }
 
-// processSingleDeploymentRow only processes the row. It assumes that there is an element ready to be processed in rows.
-func (h *DBHandler) processSingleDeploymentRow(ctx context.Context, rows *sql.Rows) (*Deployment, error) {
-	var row = &DBDeployment{
-		Created:        time.Time{},
-		ReleaseVersion: nil,
-		App:            "",
-		Env:            "",
-		Metadata:       "",
-		TransformerID:  0,
-	}
-	var releaseVersion sql.NullInt64
-	//exhaustruct:ignore
-	var resultJson = DeploymentMetadata{}
-
-	err := rows.Scan(&row.Created, &releaseVersion, &row.App, &row.Env, &row.Metadata, &row.TransformerID)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return nil, nil
-		}
-		return nil, fmt.Errorf("Error scanning deployments row from DB. Error: %w\n", err)
-	}
-	if releaseVersion.Valid {
-		row.ReleaseVersion = &releaseVersion.Int64
-	}
-
-	err = json.Unmarshal(([]byte)(row.Metadata), &resultJson)
-	if err != nil {
-		return nil, fmt.Errorf("Error during json unmarshal in deployments. Error: %w. Data: %s\n", err, row.Metadata)
-	}
-
-	return &Deployment{
-		Created:       row.Created,
-		App:           row.App,
-		Env:           row.Env,
-		Version:       row.ReleaseVersion,
-		Metadata:      resultJson,
-		TransformerID: row.TransformerID,
-	}, nil
-}
-
 // Environments
 
 type DBAllEnvironments struct {
@@ -5337,96 +4636,6 @@ func (h *DBHandler) DBReadLastFailedEslEvents(ctx context.Context, tx *sql.Tx, l
 	}
 
 	return failedEsls, nil
-}
-
-type AllDeploymentsForApp struct {
-	Version     EslVersion       `json:"version"`
-	AppName     string           `json:"appName"`
-	Deployments map[string]int64 `json:"deployments"` //Maps environment name to release version
-}
-
-func (h *DBHandler) DBWriteAllDeploymentsForApp(ctx context.Context, tx *sql.Tx, prev int, appName string, environmentDeployments map[string]int64) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBWriteAllDeploymentsForApp")
-	defer span.Finish()
-
-	if h == nil {
-		return nil
-	}
-	if tx == nil {
-		return fmt.Errorf("attempting to write to the all_deployments table without a transaction")
-	}
-
-	insertQuery := h.AdaptQuery(
-		"INSERT INTO all_deployments (eslVersion, created, appName, json) VALUES (?, ?, ?, ?);",
-	)
-	now, err := h.DBReadTransactionTimestamp(ctx, tx)
-	if err != nil {
-		return fmt.Errorf("DBWriteAllDeploymentsForApp unable to get transaction timestamp: %w", err)
-	}
-	jsonDeployments, err := json.Marshal(environmentDeployments)
-	if err != nil {
-		return fmt.Errorf("could not marshall deployments for app: '%s': %v\n", appName, err)
-	}
-
-	span.SetTag("query", insertQuery)
-	_, err = tx.Exec(
-		insertQuery,
-		prev+1,
-		*now,
-		appName,
-		jsonDeployments,
-	)
-	if err != nil {
-		return fmt.Errorf("DBWriteAllDeploymentsForApp error executing query: %w", err)
-	}
-	return nil
-}
-
-type DBAllDeploymentsForAppRow struct {
-	EslVersion      EslVersion
-	Created         time.Time
-	AppName         string
-	DeploymentsJson string
-}
-
-func (h *DBHandler) processAllDeploymentRow(ctx context.Context, err error, rows *sql.Rows) (*AllDeploymentsForApp, error) {
-	if err != nil {
-		return nil, fmt.Errorf("could not query all_deployments table from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("all_deployments: row could not be closed: %v", err)
-		}
-	}(rows)
-	//exhaustruct:ignore
-	var row = &DBAllDeploymentsForAppRow{}
-	//exhaustruct:ignore
-	var deployments = &AllDeploymentsForApp{}
-	if rows.Next() {
-
-		err := rows.Scan(&row.EslVersion, &row.Created, &row.AppName, &row.DeploymentsJson)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning oldest_deployments row from DB. Error: %w\n", err)
-		}
-		err = json.Unmarshal([]byte(row.DeploymentsJson), &deployments.Deployments)
-		if err != nil {
-			return nil, fmt.Errorf("Error unmarshalling all deployments. Error: %w\n", err)
-		}
-		deployments.AppName = row.AppName
-		deployments.Version = row.EslVersion
-	} else {
-		row = nil
-		deployments = nil
-	}
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return deployments, nil
 }
 
 func (h *DBHandler) DBReadTransactionTimestamp(ctx context.Context, tx *sql.Tx) (*time.Time, error) {
